[API] Refactored the Cart model. Then, created a migration to apply the changes to the Cart model. In cart.controller.ts, updated the endpoints and restricted access to authenticated users only. In cart.service.ts, refactored the logic in all functions. Similarly, in cart-items.controller.ts, restricted all endpoints to authenticated users only, and in cart-items.service.ts, adjusted the logic in some places. Removed an unnecessary key from cart-item.dto.ts and deleted the unused cart.dto.ts file.

+++ backend/src/cart-items/cart-items.controller.ts
@@ -11,7 +11,6 @@ import { CartItemsService } from './cart-items.service';
 import { CartItemDto } from '../dto/cart-item.dto';
 import { Request } from 'express';
 import { RequestUser } from '../types';
-import { CartDto } from '../dto/cart.dto';
 
 @Controller('cart')
 export class CartItemsController {
@@ -19,47 +18,51 @@ export class CartItemsController {
 
   @Post(':cartId/item')
   async createItem(
-    @Req() req: Request & { user?: RequestUser },
+    @Req() req: Request & { user: RequestUser },
     @Param('cartId') cartId: number,
     @Body() cartItemDto: CartItemDto,
   ) {
     const token = req.headers.authorization;
-    return await this.cartItemService.createCartItem(
-      cartId,
-      cartItemDto,
-      token,
-    );
+    if (token) {
+      return await this.cartItemService.createCartItem(
+        cartId,
+        cartItemDto,
+        token,
+      );
+    }
   }
 
   @Patch(':cartId/item/:productId')
   async updateItem(
-    @Req() req: Request & { user?: RequestUser },
+    @Req() req: Request & { user: RequestUser },
     @Param('cartId') cartId: number,
     @Param('productId') productId: number,
     @Body() cartItemDto: CartItemDto,
   ) {
     const token = req.headers.authorization;
-    return await this.cartItemService.updateCartItem(
-      cartId,
-      productId,
-      cartItemDto,
-      token,
-    );
+    if (token) {
+      return await this.cartItemService.updateCartItem(
+        cartId,
+        productId,
+        cartItemDto,
+        token,
+      );
+    }
   }
 
   @Delete(':cartId/item/:productId')
   async deleteItem(
-    @Req() req: Request & { user?: RequestUser },
+    @Req() req: Request & { user: RequestUser },
     @Param('cartId') cartId: number,
     @Param('productId') productId: number,
-    @Body() cartDto: CartDto,
   ) {
     const token = req.headers.authorization;
-    return await this.cartItemService.deleteCartItem(
-      cartId,
-      productId,
-      cartDto,
-      token,
-    );
+    if (token) {
+      return await this.cartItemService.deleteCartItem(
+        cartId,
+        productId,
+        token,
+      );
+    }
   }
 }
